SpeedDial추가

+++ lib/home_page.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_speed_dial/flutter_speed_dial.dart';
 import 'package:flutter_tabbar_exam/provider/firebase_item_provider.dart';
+import 'package:flutter_tabbar_exam/widget/fab_speed_dial.dart';
 
 class MyHomePage extends ConsumerStatefulWidget {
   const MyHomePage({super.key});
@@ -34,79 +35,6 @@ class _MyHomePageState extends ConsumerState<MyHomePage> with TickerProviderStat
     return Scaffold(
       appBar: AppBar(
         title: const Text('TabBar Example'),
-        actions: [
-          TextButton(
-            onPressed: () async {
-              showDialog(
-                context: context,
-                builder: (context) {
-                  return AlertDialog(
-                    title: const Text("Add Tab"),
-                    content: TextField(
-                      controller: textController,
-                      autofocus: true,
-                      decoration: const InputDecoration(
-                        hintText: "Enter tab name",
-                      ),
-                    ),
-                    actions: [
-                      TextButton(
-                        onPressed: () async {
-                          if (textController.text.isEmpty) {
-                            return;
-                          }
-
-                          final docRef = db.collection("category").doc(textController.text);
-
-                          // 문서 존재 여부 확인
-                          final docSnapshot = await docRef.get();
-
-                          if(docSnapshot.exists) {
-                            showDialog(
-                              context: context,
-                              builder: (context) {
-                                return AlertDialog(
-                                  title: const Text("Duplicate Tab"),
-                                  content: const Text("This tab already exists."),
-                                  actions: [
-                                    TextButton(
-                                      onPressed: () {
-                                        textController.clear();
-                                        Navigator.pop(context);
-                                      },
-                                      child: const Text("OK"),
-                                    ),
-                                  ],
-                                );
-                              },
-                            );
-                          } else {
-                            await docRef.set({
-                              "categoryName": textController.text,
-                              "createdAt": FieldValue.serverTimestamp(),
-                            });
-                            textController.clear();
-                            Navigator.pop(context);
-                          }
-
-                        },
-                        child: const Text("Add"),
-                      ),
-                      TextButton(
-                        onPressed: () {
-                          Navigator.pop(context);
-                          textController.clear();
-                        },
-                        child: const Text("Cancel"),
-                      ),
-                    ],
-                  );
-                },
-              );
-            },
-            child: const Text("탭 추가"),
-          ),
-        ],
       ),
       body: StreamBuilder<QuerySnapshot>(
         stream: db.collection("category").orderBy("createdAt", descending: false).snapshots(),
@@ -118,40 +46,51 @@ class _MyHomePageState extends ConsumerState<MyHomePage> with TickerProviderStat
             return const Center(child: Text("No tabs available."));
           }
 
-          // Firestore 데이터로 탭 목록 구성
-          List<String> tabList = snapshot.data!.docs.map((doc) {
-            return doc["categoryName"] as String;
-          }).toList();
+          // Firestore 데이터를 기반으로 탭 생성
+          final tabs = snapshot.data!.docs;
 
-
-          if (tabController == null || tabController!.length != tabList.length) {
-            tabController?.dispose();
-            tabController = TabController(length: tabList.length, vsync: this)
+          // TabController 초기화 및 기존 상태 유지
+          if (tabController == null) {
+            tabController = TabController(length: tabs.length, vsync: this)
               ..addListener(() {
                 if (tabController!.indexIsChanging) {
                   ref.read(currentTabIndexProvider.notifier).state = tabController!.index;
                 }
               });
-
+          } else if (tabController!.length != tabs.length) {
+            final previousIndex = tabController!.index;
+            tabController!.dispose();
+            tabController = TabController(
+              length: tabs.length,
+              vsync: this,
+              initialIndex: (previousIndex < tabs.length) ? previousIndex : 0,
+            )..addListener(() {
+              if (tabController!.indexIsChanging) {
+                ref.read(currentTabIndexProvider.notifier).state = tabController!.index;
+              }
+            });
           }
+
+          print(tabController!.index);
+          print(currentTabIndex);
 
           return Column(
             children: [
               TabBar(
                 controller: tabController,
-                tabs: tabList.map((tabTitle) => Tab(text: tabTitle)).toList(),
+                tabs: tabs.map((tabTitle) => Tab(text: tabTitle["categoryName"])).toList(),
                 isScrollable: true,
               ),
               Expanded(
                 child: TabBarView(
                   controller: tabController,
-                  children: tabList.map((tabTitle) {
+                  children: tabs.map((tabTitle) {
                     return Column(
                       children: [
                         StreamBuilder<QuerySnapshot>(
                           stream: db
                               .collection("category")
-                              .doc(currentCategoryValue.value!.docs[currentTabIndex].id)
+                              .doc(tabTitle.id)
                               .collection("food")
                               .snapshots(),
                           builder: (context, snapshot) {
@@ -173,8 +112,37 @@ class _MyHomePageState extends ConsumerState<MyHomePage> with TickerProviderStat
                                     subtitle: Text("${foodItems[index]["가격"]}원"),
                                     trailing: IconButton(
                                       onPressed: () {
-                                        db.collection("category").doc(currentCategoryValue.value!.docs[currentTabIndex].id)
-                                            .collection("food").doc(foodItems[index].id).delete();
+                                        showDialog(
+                                          context: context,
+                                          builder: (context) {
+                                            return AlertDialog(
+                                              title: Text("Delete ${foodItems[index]["제품 명"]}"),
+                                              content: Text("Are you sure you want to delete this item?"),
+                                              actions: [
+                                                TextButton(
+                                                  onPressed: () {
+                                                    print(tabs[currentTabIndex].id);
+                                                    print(currentCategoryValue.value!.docs[currentTabIndex].id);
+                                                    print(tabController!.index);
+                                                    print(tabs[tabController!.index].id);
+                                                    print(tabs[currentTabIndex].id);
+                                                    print(tabTitle.id);
+                                                    Navigator.pop(context);
+                                                  },
+                                                  child: Text("Cancel"),
+                                                ),
+                                                TextButton(
+                                                  onPressed: () {
+                                                    Navigator.pop(context);
+                                                    db.collection("category").doc(currentCategoryValue.value!.docs[currentTabIndex].id)
+                                                        .collection("food").doc(foodItems[index].id).delete();
+                                                  },
+                                                  child: Text("Delete"),
+                                                ),
+                                              ],
+                                            );
+                                          },
+                                        );
 
                                         print(foodItems[index].id);
                                         print(currentCategoryValue.value!.docs[currentTabIndex].id);
@@ -200,127 +168,7 @@ class _MyHomePageState extends ConsumerState<MyHomePage> with TickerProviderStat
           );
         },
       ),
-      floatingActionButton: Row(
-        mainAxisAlignment: MainAxisAlignment.end,
-        children: [
-          FloatingActionButton(
-            onPressed: () {
-
-              final documentId = currentCategoryValue.value!.docs[currentTabIndex].id;
-              print(currentCategoryValue.value!.docs[tabController!.index].id);
-              print('현재 선택된 Tab: ${tabController!.index}');
-              print(currentTabIndex);
-              print(documentId);
-
-              showDialog(
-                context: context,
-                builder: (context) {
-                  return AlertDialog(
-                    title: const Text("음식 입력"),
-                    content: Column(
-                      mainAxisSize: MainAxisSize.min,
-                      children: [
-                        TextField(
-                          controller: textController,
-                          autofocus: true,
-                          decoration: const InputDecoration(
-                            labelText: "음식명",
-                          ),
-                        ),
-                        TextField(
-                          controller: priceController,
-                          decoration: const InputDecoration(
-                            labelText: "가격",
-                          ),
-                        ),
-                      ],
-                    ),
-                    actions: [
-                      TextButton(
-                        onPressed: () async {
-
-                          if (textController.text.isEmpty || priceController.text.isEmpty) {
-                            return;
-                          }
-
-                          final docRef = db.collection("category").doc(documentId).collection("food").doc(textController.text);
-
-                          final docSnapshot = await docRef.get();
-                          print(docSnapshot.exists);
-
-                          if(docSnapshot.exists) {
-                            showDialog(
-                              context: context,
-                              builder: (context) {
-                                return AlertDialog(
-                                  title: const Text("Duplicate item"),
-                                  content: const Text("This item already exists."),
-                                  actions: [
-                                    TextButton(
-                                      onPressed: () {
-                                        textController.clear();
-                                        priceController.clear();
-                                        Navigator.pop(context);
-                                      },
-                                      child: const Text("OK"),
-                                    ),
-                                  ],
-                                );
-                              },
-                            );
-                          } else {
-                            await docRef.set({
-                              "제품 명": textController.text,
-                              "가격": priceController.text,
-                            });
-
-                            textController.clear();
-                            priceController.clear();
-                            Navigator.pop(context);
-                          }
-
-                        },
-                        child: const Text("입력"),
-                      ),
-                      TextButton(
-                        onPressed: () {
-                          textController.clear();
-                          priceController.clear();
-                          Navigator.pop(context);
-                        },
-                        child: const Text("취소"),
-                      ),
-                    ],
-                  );
-                },
-              );
-
-            },
-            child: const Icon(Icons.edit),
-          ),
-          SpeedDial(
-            animatedIcon: AnimatedIcons.menu_close,
-            overlayOpacity: 0.5,
-            children: [
-              SpeedDialChild(
-                child: const Icon(Icons.add),
-                label: 'Add Item',
-                onTap: () => print('Add Item tapped'),
-              ),
-              SpeedDialChild(
-                child: const Icon(Icons.edit),
-                label: 'Edit Item',
-                onTap: () => print('Edit Item tapped'),
-              ),
-              SpeedDialChild(
-                child: const Icon(Icons.delete),
-                label: 'Delete Item',
-                onTap: () => print('Delete Item tapped'),
-              ),
-            ],
-          ),
-        ],
-      ),
+      floatingActionButton: FabSpeedDial(),
     );
   }
 }
